Fix build breaks on mac/linux
