+++ config/config.go
@@ -4,10 +4,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"os/user"
 	"path"
-	"runtime"
-	"syscall"
 
 	yaml "gopkg.in/yaml.v2"
 )
@@ -102,30 +99,6 @@ func createConfigPath() error {
 		return err
 	}
 	return os.MkdirAll(path, 0700)
-}
-
-func getHomeDir() (string, error) {
-	// TODO: This is a workaround for user.Current being
-	// very slow on a domain joined PC that is not connected
-	// to the domain.
-	if runtime.GOOS == "windows"  {
-		t, e := syscall.OpenCurrentProcessToken()
-		if e != nil {
-			return "", e
-		}
-		defer t.Close()
-		dir, e := t.GetUserProfileDirectory()
-		if e != nil {
-			return "", e
-		}
-		return dir, nil
-	}
-
-	usr, err := user.Current()
-	if err != nil {
-		return "", err
-	}
-	return usr.HomeDir, nil
 }
 
 // GetConfigFilePath gets the full path to the given config file name.
