+++ config/config_notwin.go
@@ -1,0 +1,15 @@
+// +build !windows
+
+package config
+
+import (
+	"os/user"
+)
+
+func getHomeDir() (string, error) {
+	usr, err := user.Current()
+	if err != nil {
+		return "", err
+	}
+	return usr.HomeDir, nil
+}
