+++ service/debugger/debugger_linux.go
@@ -31,5 +31,5 @@ func attachErrorMessage(pid int, err error) error {
 }
 
 func stopProcess(pid int) error {
-	return sys.Kill(d.ProcessPid(), sys.SIGSTOP)
+	return sys.Kill(pid, sys.SIGSTOP)
 }
